// src/common/enums/ProgLangsEnum.ts
export enum ProgLangsEnum {
  JavaScript = 'JavaScript',
  TypeScript = 'TypeScript',
  Python = 'Python',
  Go = 'Go',
  C = 'C',
  'C++' = 'C++',
  'C#' = 'C#',
  Ruby = 'Ruby',
  Java = 'Java',
  Lua = 'Lua',
  Lisp = 'Lisp',
  Haskell = 'Haskell',
  Rust = 'Rust',
  Kotlin = 'Kotlin',
  Php = 'Php',
  Julia = 'Julia',
}